[Silver III] Title: 소수 구하기, Time: 480 ms, Memory: 20920 KB -BaekjoonHub

+++ 백준/Silver/1929. 소수 구하기/소수 구하기.java
@@ -22,7 +22,7 @@ public class Main {
     }
 
     // 소수 판별
-    private static boolean isPrime(long n) {
+    private static boolean isPrime(int n) {
         // 0, 1
         if (n < 2) {
             return false;
